+++ CHANGES.rst
@@ -8,6 +8,8 @@ Changelog
   [sgeulette]
 - Added `setup.load_xml_tool_only_from_package` to load only main tool xml file.
   [sgeulette]
+- Added `setup.test_remove_gs_step` to remove a generic setup step.
+  [sgeulette]
 
 0.80 (2023-12-11)
 -----------------